sonostool/requirements.txt
docopt
requests
pycryptodome
